+++ test/simple_test.dart
@@ -9,7 +9,7 @@ void main() {
     });
 
     test('android', () {
-      WebsafePlatform.override(_MockWebsafePlatform(android: true));
+      WebsafePlatform.override(WebsafePlatform.mock(android: true));
       var websafePlatform = WebsafePlatform();
       expect(true, websafePlatform.isAndroid());
       expect(false, websafePlatform.isFuchsia());
@@ -18,10 +18,25 @@ void main() {
       expect(false, websafePlatform.isMacOS());
       expect(false, websafePlatform.isWeb());
       expect(false, websafePlatform.isWindows());
+
+      expect('android', WebsafePlatform.name(websafePlatform));
+    });
+    test('fuchsia', () {
+      WebsafePlatform.override(WebsafePlatform.mock(fuchsia: true));
+      var websafePlatform = WebsafePlatform();
+      expect(false, websafePlatform.isAndroid());
+      expect(true, websafePlatform.isFuchsia());
+      expect(false, websafePlatform.isIOS());
+      expect(false, websafePlatform.isLinux());
+      expect(false, websafePlatform.isMacOS());
+      expect(false, websafePlatform.isWeb());
+      expect(false, websafePlatform.isWindows());
+
+      expect('fuchsia', WebsafePlatform.name(websafePlatform));
     });
 
     test('ios', () {
-      WebsafePlatform.override(_MockWebsafePlatform(ios: true));
+      WebsafePlatform.override(WebsafePlatform.mock(ios: true));
       var websafePlatform = WebsafePlatform();
       expect(false, websafePlatform.isAndroid());
       expect(false, websafePlatform.isFuchsia());
@@ -30,10 +45,40 @@ void main() {
       expect(false, websafePlatform.isMacOS());
       expect(false, websafePlatform.isWeb());
       expect(false, websafePlatform.isWindows());
+
+      expect('ios', WebsafePlatform.name(websafePlatform));
+    });
+
+    test('linux', () {
+      WebsafePlatform.override(WebsafePlatform.mock(linux: true));
+      var websafePlatform = WebsafePlatform();
+      expect(false, websafePlatform.isAndroid());
+      expect(false, websafePlatform.isFuchsia());
+      expect(false, websafePlatform.isIOS());
+      expect(true, websafePlatform.isLinux());
+      expect(false, websafePlatform.isMacOS());
+      expect(false, websafePlatform.isWeb());
+      expect(false, websafePlatform.isWindows());
+
+      expect('linux', WebsafePlatform.name(websafePlatform));
+    });
+
+    test('macos', () {
+      WebsafePlatform.override(WebsafePlatform.mock(macos: true));
+      var websafePlatform = WebsafePlatform();
+      expect(false, websafePlatform.isAndroid());
+      expect(false, websafePlatform.isFuchsia());
+      expect(false, websafePlatform.isIOS());
+      expect(false, websafePlatform.isLinux());
+      expect(true, websafePlatform.isMacOS());
+      expect(false, websafePlatform.isWeb());
+      expect(false, websafePlatform.isWindows());
+
+      expect('macos', WebsafePlatform.name(websafePlatform));
     });
 
     test('web', () {
-      WebsafePlatform.override(_MockWebsafePlatform(web: true));
+      WebsafePlatform.override(WebsafePlatform.mock(web: true));
       var websafePlatform = WebsafePlatform();
       expect(false, websafePlatform.isAndroid());
       expect(false, websafePlatform.isFuchsia());
@@ -42,47 +87,22 @@ void main() {
       expect(false, websafePlatform.isMacOS());
       expect(true, websafePlatform.isWeb());
       expect(false, websafePlatform.isWindows());
+
+      expect('web', WebsafePlatform.name(websafePlatform));
+    });
+
+    test('windows', () {
+      WebsafePlatform.override(WebsafePlatform.mock(windows: true));
+      var websafePlatform = WebsafePlatform();
+      expect(false, websafePlatform.isAndroid());
+      expect(false, websafePlatform.isFuchsia());
+      expect(false, websafePlatform.isIOS());
+      expect(false, websafePlatform.isLinux());
+      expect(false, websafePlatform.isMacOS());
+      expect(false, websafePlatform.isWeb());
+      expect(true, websafePlatform.isWindows());
+
+      expect('windows', WebsafePlatform.name(websafePlatform));
     });
   });
-}
-
-class _MockWebsafePlatform implements WebsafePlatform {
-  _MockWebsafePlatform({
-    this.android = false,
-    this.fuchsia = false,
-    this.ios = false,
-    this.linux = false,
-    this.macos = false,
-    this.web = false,
-    this.windows = false,
-  });
-
-  final bool android;
-  final bool fuchsia;
-  final bool ios;
-  final bool linux;
-  final bool macos;
-  final bool web;
-  final bool windows;
-
-  @override
-  bool isAndroid() => android;
-
-  @override
-  bool isFuchsia() => fuchsia;
-
-  @override
-  bool isIOS() => ios;
-
-  @override
-  bool isLinux() => linux;
-
-  @override
-  bool isMacOS() => macos;
-
-  @override
-  bool isWeb() => web;
-
-  @override
-  bool isWindows() => windows;
 }